scripts/requirements.txt
python4yahdlc==1.3.5
autopep8==2.0.4
websockets==11.0.3